+++ node/src/main/kotlin/net/corda/node/services/config/ConfigUtilities.kt
@@ -7,12 +7,8 @@ import com.typesafe.config.ConfigRenderOptions
 import net.corda.core.crypto.Crypto
 import net.corda.core.crypto.SignatureScheme
 import net.corda.core.identity.CordaX500Name
-import net.corda.core.internal.copyTo
-import net.corda.core.internal.createDirectories
-import net.corda.core.internal.div
-import net.corda.core.internal.exists
+import net.corda.core.internal.*
 import net.corda.core.utilities.loggerFor
-import net.corda.core.utilities.x500Name
 import net.corda.node.utilities.*
 import net.corda.nodeapi.config.SSLConfiguration
 import org.bouncycastle.asn1.x509.GeneralName
